Refactored keyword adding and display to properly modularize the components. This will allow the keyword-add and keyword-list components to be independently positioned.
