+++ app/components/snapshot.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Fluxxor from 'fluxxor'
-import KeywordList from './keywordlist'
+import ActiveKeywordList from './activekeywordlist'
+import AddKeyword from './addkeyword'
 import PublisherList from './publisherlist'
 import Slider from './slider'
 import { Button } from 'react-bootstrap'
@@ -20,7 +21,8 @@ var options = {
   barStrokeWidth : 2,
   barValueSpacing : 5,
   barDatasetSpacing : 1,
-  legendTemplate : "<ul class=\"<%=name.toLowerCase()%>-legend\"><% for (var i=0; i<datasets.length; i++){%><li><span style=\"background-color:<%=datasets[i].fillColor%>\"></span><%if(datasets[i].label){%><%=datasets[i].label%><%}%></li><%}%></ul>"
+  // legendTemplate : "<ul class=\"<%=name.toLowerCase()%>-legend\"><% for (var i=0; i<datasets.length; i++){%><li><span style=\"background-color:<%=datasets[i].fillColor%>\"></span><%if(datasets[i].label){%><%=datasets[i].label%><%}%></li><%}%></ul>"
+  // legendTemplate : "<ul class=\"<%=name.toLowerCase()%>-legend\"><% for (var i=0; i<datasets.length; i++){%><li><span style=\"background-color:<%=datasets[i].fillColor%>\"></span><%if(datasets[i].label){%><%=datasets[i].label%><%}%></li><%}%></ul>"
 };
 
 export default React.createClass({
@@ -51,10 +53,12 @@ export default React.createClass({
           </div>
           <i onClick={this.toggleHidden} className="fa fa-2x fa-cog chart-options"></i>
           <div className={(this.props.chartParams.hiddenSettings ? 'hidden ' : '') + "chart-menu"}>
-            <KeywordList chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
-            <PublisherList chartID={this.props.chartParams.chartID} 
-                           className={'publisher-list'} 
-                           list={this.props.publisherList} 
+            <h5>Keywords</h5>
+              <AddKeyword chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
+              <ActiveKeywordList chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
+            <PublisherList chartID={this.props.chartParams.chartID}
+                           className={'publisher-list'}
+                           list={this.props.publisherList}
                            activelist={this.props.chartParams.publishers}
                            legend={true}/>
             <Button className="delete" bsStyle="danger">Delete Chart</Button>
